requirements.txt
python-telegram-bot==20.8
pytz==2024.1
python-dotenv==1.0.1
aiohttp==3.9.3
asyncio==3.4.3